# Reduce state machine creation