--- a/Fluid/Ast/OutputStatement.cs
+++ b/Fluid/Ast/OutputStatement.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
+using Fluid.Values;
 
 namespace Fluid.Ast
 {
@@ -16,15 +17,29 @@ namespace Fluid.Ast
 
         public IList<FilterExpression> Filters { get ; }
 
-        public override async ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
+        public override ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {
+            static async ValueTask<Completion> Awaited(
+                ValueTask<FluidValue> t,
+                TextWriter w,
+                TextEncoder enc,
+                TemplateContext ctx)
+            {
+                var value = await t;
+                value.WriteTo(w, enc, ctx.CultureInfo);
+                return Completion.Normal;
+            }
+
             context.IncrementSteps();
 
-            var value = await Expression.EvaluateAsync(context);
+            var task = Expression.EvaluateAsync(context);
+            if (task.IsCompletedSuccessfully)
+            {
+                task.Result.WriteTo(writer, encoder, context.CultureInfo);
+                return new ValueTask<Completion>(Completion.Normal);
+            }
 
-            value.WriteTo(writer, encoder, context.CultureInfo);
-
-            return Completion.Normal;
+            return Awaited(task, writer, encoder, context);
         }
     }
 }