--- a/Fluid/StringSegmentExtensions.cs
+++ b/Fluid/StringSegmentExtensions.cs
@@ -1,17 +1,19 @@
-﻿using Microsoft.Extensions.Primitives;
+﻿using System.Runtime.CompilerServices;
 
 namespace Microsoft.Extensions.Primitives
 {
     public static class StringSegmentExtensions
     {
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static char Index(this StringSegment segment, int index)
         {
+            var indexToUse = segment.Offset + index;
             if (index < 0)
             {
-                return segment.Buffer[segment.Offset + segment.Length + index];
+                indexToUse += segment.Length;
             }
 
-            return segment.Buffer[segment.Offset + index];
+            return segment.Buffer[indexToUse];
         }
     }
 }