… Fluid/Values/FactoryValue.cs
… Fluid/Values/NumberValue.cs
… Fluid/Values/ObjectValue.cs
… Fluid/Values/StringValue.cs
… appveyor.yml