Making code a bit uglier, but now detecting fast path where async targets are already resolved without need for `await`, this removes the state machine from call site. When looking at OC's traces it's a deeeep call chain because of state machinery, this PR paves way for tweaking OC behavior.

Profiled against the benchmark test case, probably could later check binary expressions etc, don't know how often they are used. This basically removes all state machines in benchmark test case's liquid template except for `ForStatement`.

* use idiomatic ExceptionHelper
* prepare some properties and helpers for better size prediction
* add netstandard 2.1 target
* upgrade support libs

## Fluid.Benchmarks.FluidBenchmarks

| **Diff**|Method|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------:|-------:|-------:|-------:|
| Old |Parse|35.85 us|7.0190|-|-|28.81 KB|
| **New** |	| **37.19 us (+4%)** | **7.0190 (0%)** | **-** | **-** | **28.81 KB (0%)** |
| Old |Render|1,158.34 us|58.5938|29.2969|29.2969|224.52 KB|
| **New** |	| **946.49 us (-18%)** | **58.5938 (0%)** | **57.6172 (+97%)** | **29.2969 (0%)** | **224.52 KB (0%)** |
| Old |ParseAndRender|1,282.51 us|58.5938|29.2969|29.2969|278.01 KB|
| **New** |	| **1,069.47 us (-17%)** | **58.5938 (0%)** | **29.2969 (0%)** | **29.2969 (0%)** | **278.01 KB (0%)** |