@lahma 2019-09-19
## Stack before

![image](https://user-images.githubusercontent.com/171892/65252085-06614900-db01-11e9-824f-91004dc1bb18.png)

## Stack after

![image](https://user-images.githubusercontent.com/171892/65252095-0bbe9380-db01-11e9-8354-3be0a06f87c2.png)