@sebastienros 2019-09-19
Scriban's approach is to have the code be non-async by default, and use code generation to generate the async counterparts. And you can see in his benchmarks that sync is way faster than async. I also moved most of the things to valuetask which improved the perf a lot. So your changes make sense.