@lahma 2019-09-19
Generic async makes things uniform, but in templates a few items are I/O bound, this brings back speed with cost of maintainability/readability.